requirements.txt
Django==3.2.7
numpy==1.21.6
pyenchant==3.2.2
